% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-go.R
\docType{data}
\name{go_cc}
\alias{go_cc}
\title{Example GO cellular component graph}
\format{
An S4 \code{GO} object containing term metadata and the
parent–child DAG structure.
}
\source{
Gene Ontology
}
\description{
A Gene Ontology (GO) graph restricted to the Cellular Component (CC)
ontology. This dataset is intended for use in examples and documentation.
}
\details{
The object is of class \code{GO} and contains GO term metadata and
parent–child relationships for the CC ontology.


The graph was created with:

\preformatted{
load_go(ont = "CC")
}
}
\examples{
data("go_cc", package = "GOcontext")
go_cc

}
\keyword{internal}
