% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-go.R
\docType{data}
\name{go_cc_ecoli}
\alias{go_cc_ecoli}
\title{Example GO CC graph with E. coli gene mappings}
\format{
An S4 \code{GO} object with an attached GO-to-gene mapping.
}
\source{
Gene Ontology and org.EcK12.eg.db
}
\description{
A Gene Ontology Cellular Component graph with organism mappings
attached for \emph{Escherichia coli} K-12 using the
\code{org.EcK12.eg.db} annotation package.
}
\details{
This dataset is intended for use in runnable examples demonstrating
functions such as \code{\link{attach_org}}, \code{\link{filter_mapped}},
\code{\link{as_term2gene}}, and \code{\link{as_term2name}}.


The object was generated with:

\preformatted{
go_cc <- load_go(ont = "CC")
go_cc_ecoli <- attach_org(
    go = go_cc,
    OrgDb = org.EcK12.eg.db::org.EcK12.eg.db,
    keytype = "ENTREZID"
)
}
}
\examples{
data("go_cc_ecoli", package = "GOcontext")
go_cc_ecoli

}
\keyword{internal}
