% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_go.R
\name{subset_go}
\alias{subset_go}
\title{Restrict a GO graph by keeping or excluding selected branches}
\usage{
subset_go(go, ids, mode = c("keep", "exclude"))
}
\arguments{
\item{go}{A \code{GO} or \code{GOSubgraph} object.}

\item{ids}{\code{character()} Vector of GO IDs used as subsetting seeds.}

\item{mode}{\code{character(1)} One of \code{"keep"} or
\code{"exclude"}.}
}
\value{
A \code{GOSubgraph} object.
}
\description{
Given one or more GO IDs, induce a subgraph by either keeping those terms
and all their descendants, or excluding those terms and all their
descendants, within the current graph.
}
\examples{
data("go_cc_ecoli", package = "GOcontext")

go_cc_ecoli_keep <- subset_go(
    go = go_cc_ecoli,
    ids = "GO:0009279",
    mode = "keep"
)
go_cc_ecoli_keep

go_cc_ecoli_exclude <- subset_go(
    go = go_cc_ecoli,
    ids = "GO:0009279",
    mode = "exclude"
)
go_cc_ecoli_exclude

}
